+++ src/test/scala/satify/update/converters/CNFValidator.scala
@@ -1,10 +1,9 @@
-package satify.ExpressionUtilsTest
+package satify.update.converters
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 import satify.model.expression.Expression
 import satify.model.expression.Expression.*
-import satify.model.{CNF, Variable}
 import satify.update.converters.TseitinTransformation.isCNF
 
 class CNFValidator extends AnyFlatSpec with Matchers:
