+++ src/test/scala/satify/update/converters/VariableSubstitutionTest.scala
@@ -1,10 +1,10 @@
-package satify.ExpressionUtilsTest
+package satify.update.converters
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import satify.model.expression.Expression.*
 import satify.model.Variable
 import satify.model.expression.Expression
+import satify.model.expression.Expression.*
 import satify.update.converters.TseitinTransformation.symbolsReplace
 
 class VariableSubstitutionTest extends AnyFlatSpec with Matchers:
