+++ src/test/scala/satify/update/solver/dpll/CNFSimplificationTest.scala
@@ -1,12 +1,12 @@
-package satify.DPLLTest
+package satify.update.solver.dpll
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import satify.update.solver.dpll.CNFSimplification.*
-import satify.model.{CNF, Variable}
-import satify.model.CNF.*
 import satify.model.Bool.{False, True}
+import satify.model.CNF.*
 import satify.model.dpll.{Constraint, PartialModel}
+import satify.model.{CNF, Variable}
+import satify.update.solver.dpll.CNFSimplification.*
 
 class CNFSimplificationTest extends AnyFlatSpec with Matchers:
