+++ src/test/scala/satify/update/solver/dpll/ConflictIdentificationTest.scala
@@ -1,13 +1,13 @@
-package satify.DPLLTest
+package satify.update.solver.dpll
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import satify.model.{CNF, Variable}
-import satify.model.CNF.{And, Not, Or, Symbol}
 import satify.model.Bool.{False, True}
+import satify.model.CNF.{And, Not, Or, Symbol}
 import satify.model.dpll.Constraint
-import satify.update.solver.dpll.ConflictIdentification.isUnsat
+import satify.model.{CNF, Variable}
 import satify.update.solver.dpll.CNFSimplification.simplifyCnf
+import satify.update.solver.dpll.ConflictIdentification.isUnsat
 
 class ConflictIdentificationTest extends AnyFlatSpec with Matchers:
