+++ src/test/scala/satify/update/solver/dpll/DPLLTest.scala
@@ -1,20 +1,21 @@
-package satify.DPLLTest
+package satify.update.solver.dpll
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.matchers.should.Matchers.should
-import satify.model.{CNF, Variable}
-import satify.model.dpll.DecisionTree.*
-import satify.model.CNF.*
 import satify.model.Bool.{False, True}
-import satify.update.solver.dpll.CNFSimplification.*
-import satify.update.solver.dpll.PartialModelUtils.*
-import satify.model.dpll.OrderedSeq.given_Ordering_Variable
+import satify.model.CNF.*
+import satify.model.dpll.DecisionTree.*
 import satify.model.dpll.OrderedSeq.*
 import satify.model.dpll.{Decision, DecisionTree, PartialModel}
+import satify.model.{CNF, Variable}
+import satify.update.solver.dpll.CNFSimplification.*
 import satify.update.solver.dpll.DPLL.*
+import satify.update.solver.dpll.PartialModelUtils.*
 
 class DPLLTest extends AnyFlatSpec with Matchers:
+
+  import satify.model.dpll.OrderedSeq.given_Ordering_Variable
 
   val varA: Variable = Variable("a")
   val varB: Variable = Variable("b")
