+++ src/test/scala/satify/update/solver/dpll/PartialModelUtilsTest.scala
@@ -1,13 +1,13 @@
-package satify.DPLLTest
+package satify.update.solver.dpll
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import satify.update.solver.dpll.PartialModelUtils.*
 import satify.model.CNF.*
-import satify.model.dpll.OrderedSeq.{seq, given_Ordering_Variable}
+import satify.model.dpll.OrderedSeq.{given_Ordering_Variable, seq}
 import satify.model.dpll.{Constraint, Decision, PartialModel}
 import satify.model.{CNF, Variable}
 import satify.update.solver.dpll.DPLL.dpll
+import satify.update.solver.dpll.PartialModelUtils.*
 
 class PartialModelUtilsTest extends AnyFlatSpec with Matchers:
